restruct project
