+++ Social network page/social_network_page/lib/model/model.dart
@@ -1,0 +1,15 @@
+class Profile {
+  final String imagePath;
+  final String firstName;
+  final String lastName;
+  final String dateOfBirth;
+  final String about;
+
+  Profile({
+    required this.imagePath,
+    required this.firstName,
+    required this.lastName,
+    required this.dateOfBirth,
+    required this.about,
+  });
+}
